requirements.txt
appdirs==1.4.0
bcrypt==3.1.2
Cerberus==0.9.2
cffi==1.9.1
click==6.7
Eve==0.7
Events==0.2.2
Flask==0.12
Flask-API==0.6.9
Flask-JWT==0.3.2
Flask-PyMongo==0.4.1
itsdangerous==0.24
Jinja2==2.9.5
MarkupSafe==0.23
packaging==16.8
pycparser==2.17
PyJWT==1.4.2
pymongo==3.4.0
pyparsing==2.1.10
simplejson==3.10.0
six==1.10.0
Werkzeug==0.11.14
